requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
psycopg2==2.8.6
python-telegram-bot==13.3
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2
SQLAlchemy==1.3.23
tornado==6.1
tzlocal==2.1
urllib3==1.26.3
Werkzeug==1.0.1
